+++ clis/xiaohongshu/unfollow.js
@@ -13,6 +13,7 @@
 import { cli, Strategy } from '@jackwener/opencli/registry';
 import { ArgumentError, AuthRequiredError, CliError, CommandExecutionError } from '@jackwener/opencli/errors';
 import { normalizeXhsUserId } from './user-helpers.js';
+import { unwrapEvaluateResult } from './shared.js';
 
 const PROFILE_SETTLE_MS = 2500;
 const MODAL_SETTLE_MS = 1500;
@@ -22,13 +23,6 @@ const USER_ID_RE = /^[a-zA-Z0-9]{8,32}$/;
 function isXiaohongshuHost(hostname) {
     const host = String(hostname || '').toLowerCase();
     return host === 'xiaohongshu.com' || host.endsWith('.xiaohongshu.com');
-}
-
-function unwrapEvaluateResult(payload) {
-    if (payload && typeof payload === 'object' && 'session' in payload && 'data' in payload) {
-        return payload.data;
-    }
-    return payload;
 }
 
 function requireActionResult(payload, context) {
